Create folder 'utils' and move 'util.js' into it
